refactor: consolidate DeprecationError into @taquito/core
re #1994
